messagesViewからfetchMessages*を分離
