+++ src/store/ui/mainView.ts
@@ -148,7 +148,7 @@ const useMainViewStorePinia = defineStore('ui/mainView', () => {
       channelId,
       entryMessageId
     }
-    messagesView.resetViewState()
+    messagesView.resetRenderedContent()
 
     // 通常のチャンネルは最後に開いたチャンネルとして保持
     const channelPath = channelIdToPathString(
@@ -173,7 +173,7 @@ const useMainViewStorePinia = defineStore('ui/mainView', () => {
       userName,
       entryMessageId
     }
-    messagesView.resetViewState()
+    messagesView.resetRenderedContent()
   }
 
   const changePrimaryViewToClip = ({
@@ -185,7 +185,7 @@ const useMainViewStorePinia = defineStore('ui/mainView', () => {
       type: 'clips',
       clipFolderId
     }
-    messagesView.resetViewState()
+    messagesView.resetRenderedContent()
   }
 
   return {
